Refactored the retrieval of sessions.
By doing this, we can retrieve workshops as well, and we can now add the
methods to retrieve speakers too. The latter is not done yet, but the
boilerplate for this is created. We only have to create the parsers, but
I'll do that later.

I also have to clean up the code to make the linters happy.

Fixes #15